Refactor sidebar handling and enable widget js/css
There were getting to be quite a few sidebar-related partials, so they
all got moved to a `sidebar/` folder.

Sidebar widgets can now specify a single partial to be included at the
end of the `<head>` element (i.e. CSS) and/or a single partial to be included at
the end of the `<body>` (i.e. JavaScript). Bundled sidebar widgets have
their own CSS and JS code (except that in  `styles.css`) refactored into
these partials and included the same way it would for any user's custom widgets.

Also fixed widgets not loading because they were testing to be enabled
the deprecated way. This detection happens elsewhere now, so it is
safely removed.