Fix nil panic when processing member social icons
Attempting to render memberboxes for people with social icons defined
caused nil errors because `"sitevar"` was not defined in the `.Scratch`.
This wasn't necessary for members anyways, since the code using it
determines if the website belongs to one or more people - this is
irrelevant for parsing one person's icons. The code now uses the user's
name as their pronoun.