Update CoinHive widget for users blocking CoinHive
Even though CoinHive tries to ensure that their `authedmine.com` domain
is not blocked by adblockers, it is blocked by uBlock Origin, if not
others. In the case that the JavaScript file is blocked, disable the
miner controls and show a message saying that the miner failed to load.

This commit also adds an option for people who don't want to run a miner
but may be willing to donate Monero directly by allowing users to click
a button and copy the wallet address to their clipboard. The site owner
has to provide the wallet address in their configuration to use this
behavior.